backend/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-cors==3.0.10
psycopg2-binary==2.9.1
python-dotenv==0.19.0
simple-salesforce==1.11.4
apscheduler==3.8.1
gunicorn==20.1.0
requests==2.26.0
sqlalchemy==1.4.23
